inline quest completion logic
