+++ web/components/quests-or-streak.tsx
@@ -6,7 +6,6 @@ import { Row } from 'web/components/layout/row'
 import { formatMoney } from 'common/util/format'
 import { Modal } from 'web/components/layout/modal'
 import { Col } from 'web/components/layout/col'
-import { sum } from 'lodash'
 import { dailyStatsClass } from 'web/components/daily-stats'
 import { InfoTooltip } from './widgets/info-tooltip'
 import { hasCompletedStreakToday } from 'web/components/profile/betting-streak-modal'
@@ -16,11 +15,8 @@ import {
   BETTING_STREAK_BONUS_AMOUNT,
   BETTING_STREAK_BONUS_MAX,
 } from 'common/economy'
-import { QUEST_DETAILS, QUEST_TYPES } from 'common/quest'
-import { filterDefined } from 'common/util/array'
-import { getQuestScores } from 'common/supabase/set-scores'
+import { QUEST_DETAILS } from 'common/quest'
 import { useQuestStatus } from 'web/hooks/use-quest-status'
-import { db } from 'web/lib/supabase/db'
 import { LoadingIndicator } from 'web/components/widgets/loading-indicator'
 import Link from 'next/link'
 import { linkClass } from './widgets/site-link'
@@ -212,50 +208,4 @@ const QuestRow = (props: {
       </Col>
     </Row>
   )
-}
-export const getQuestCompletionStatus = async (user: User) => {
-  const questTypes = QUEST_TYPES.filter(
-    (questType) => questType !== 'ARCHAEOLOGIST'
-  )
-  const questToCompletionStatus = Object.fromEntries(
-    questTypes.map((t) => [t, { requiredCount: 0, currentCount: 0 }])
-  )
-  const keys = questTypes.map((questType) => QUEST_DETAILS[questType].scoreId)
-  const scores = await getQuestScores(user.id, keys, db)
-
-  questTypes.forEach((questType) => {
-    const questData = QUEST_DETAILS[questType]
-    if (questType === 'BETTING_STREAK')
-      questToCompletionStatus[questType] = {
-        requiredCount: questData.requiredCount,
-        currentCount: hasCompletedStreakToday(user) ? 1 : 0,
-      }
-    else
-      questToCompletionStatus[questType] = {
-        requiredCount: questData.requiredCount,
-        currentCount: scores[questData.scoreId].score,
-      }
-  })
-
-  const totalQuestsCompleted = sum(
-    Object.values(questToCompletionStatus).map((v) =>
-      v.currentCount >= v.requiredCount ? 1 : 0
-    )
-  )
-  const incompleteQuestTypes = filterDefined(
-    Object.entries(questToCompletionStatus).map(([k, v]) =>
-      v.currentCount < v.requiredCount ? k : null
-    )
-  )
-
-  const totalQuests = Object.keys(questToCompletionStatus).length
-  const allQuestsComplete = totalQuestsCompleted === totalQuests
-
-  return {
-    questToCompletionStatus,
-    totalQuestsCompleted,
-    totalQuests,
-    allQuestsComplete,
-    incompleteQuestTypes,
-  }
 }
